[REMOVAL] Deprecate "Object" form component, rename to "SectionObject"
On PHP version 7.2+ the class name "Object" is reserved even if used in a namespace. Renaming the Flux "Object" component to "SectionObject" makes it compatible.

An alias is created for compatibility on PHP versions below 7.2, but any TYPO3 site installed on PHP 7.2 and using Flux, will need to use the new class name if referring to the component.

The ViewHelper API is untouched - and remains compatible. There is at least one other existing, similar case were we kept the VH names but renamed the PHP objects, which is okay since 99% of Flux users only ever use the VH API.

Credit for all research behind patch goes to @Lagerregal.

Close: #1487

# Conflicts:
#	Classes/Form/Container/Sheet.php
#	ext_localconf.php